% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigo.R
\name{extenso}
\alias{extenso}
\title{Numeros por extenso}
\usage{
extenso(numero_pre, moeda)
}
\arguments{
\item{numero_pre}{e o numero a ser inserido, que pode ser decimal, separado por "."}

\item{moeda, }{boolean, em caso de verdaderio o valor sera considerado monetario}
}
\value{
O numero por extenso
}
\description{
Converte os valores numericos em seu equivalente por escrito
}
\examples{
extenso1 <- extenso(50);
extenso2 <- extenso(50, TRUE);
}
